Corrigindo código de status HTTP e deletando usuário da pessoa.

+++ imob-web/back-end/src/api/controllers/administrator.controller.ts
@@ -1,7 +1,9 @@
 import { DeleteResult, EntityManager, getManager } from 'typeorm';
 import { Request, Response } from 'express';
 import { AdministratorEntity } from '../entities/administrator.entity';
+import { UserEntity } from '../entities/user.entity';
 import { AdministratorService } from '../services/administrator.service';
+import { UserService } from '../services/user.service';
 import { returnMessages, statusMessages } from '../../../utils/utils';
 
 export class AdministratorController {
@@ -27,21 +29,21 @@ export class AdministratorController {
             try {
                 const administratorService: AdministratorService =
                     new AdministratorService();
-    
+
                 const result: boolean =
                     administratorService.validateData(request.body);
-    
+
                 if (result) {
                     const result: boolean =
                         await administratorService.alreadyRegisterByCPF(request.body.cpf);
-    
+
                     if (!result) {
                         const administratorEntity: AdministratorEntity =
                             await administratorService.create(request.body, transaction);
-    
+
                         return response.status(201).json(administratorEntity);
                     } else {
-                        return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[3]}` });
+                        return response.status(409).json({ message: `${statusMessages[409]} ${returnMessages[3]}` });
                     }
                 } else {
                     return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[0]}` });
@@ -75,19 +77,19 @@ export class AdministratorController {
             try {
                 const administratorService: AdministratorService =
                     new AdministratorService();
-    
+
                 if (Number(request.params.id)) {
                     const result: boolean =
                         await administratorService.alreadyRegisterById(Number(request.params.id));
-    
+
                     if (result) {
                         const result: boolean =
                             administratorService.validateData(request.body);
-    
+
                         if (result) {
                             const administratorEntity: AdministratorEntity =
                                 await administratorService.update(Number(request.params.id), request.body, transaction);
-    
+
                             return response.status(200).json(administratorEntity);
                         } else {
                             return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[0]}` });
@@ -109,12 +111,35 @@ export class AdministratorController {
             try {
                 const administratorService: AdministratorService =
                     new AdministratorService();
-    
+
                 if (Number(request.params.id)) {
-                    const deleteResult: DeleteResult =
-                        await administratorService.delete(Number(request.params.id), transaction);
-    
-                    return response.status(200).json({ manager: deleteResult.affected });
+                    const administratorEntity: AdministratorEntity | undefined =
+                        await administratorService.read(Number(request.params.id));
+
+                    if (administratorEntity) {
+                        const userService: UserService =
+                            new UserService();
+
+                        const userEntity: UserEntity | undefined =
+                            await userService.findByAdministrator(Number(request.params.id));
+
+                        if (userEntity) {
+                            const userDeleteResult: DeleteResult =
+                                await userService.delete(userEntity.id, transaction);
+
+                            const administratorDeleteResult: DeleteResult =
+                                await administratorService.delete(Number(request.params.id), transaction);
+
+                            return response.status(200).json({
+                                user: userDeleteResult.affected,
+                                administrator: administratorDeleteResult.affected
+                            });
+                        } else {
+                            return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[1]}` });
+                        }
+                    } else {
+                        return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[1]}` });
+                    }
                 } else {
                     return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[2]}` });
                 }
