+++ imob-web/back-end/src/api/controllers/broker.controller.ts
@@ -1,7 +1,9 @@
 import { DeleteResult, EntityManager, getManager } from 'typeorm';
 import { Request, Response } from 'express';
 import { BrokerEntity } from '../entities/broker.entity';
+import { UserEntity } from '../entities/user.entity';
 import { BrokerService } from '../services/broker.service';
+import { UserService } from '../services/user.service';
 import { returnMessages, statusMessages } from '../../../utils/utils';
 
 export class BrokerController {
@@ -27,21 +29,21 @@ export class BrokerController {
             try {
                 const brokerService: BrokerService =
                     new BrokerService();
-    
+
                 const result: boolean =
                     brokerService.validateData(request.body);
-    
+
                 if (result) {
                     const result: boolean =
                         await brokerService.alreadyRegisterByCPF(request.body.cpf);
-    
+
                     if (!result) {
                         const brokerEntity: BrokerEntity =
                             await brokerService.create(request.body, transaction);
-    
+
                         return response.status(201).json(brokerEntity);
                     } else {
-                        return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[3]}` });
+                        return response.status(409).json({ message: `${statusMessages[409]} ${returnMessages[3]}` });
                     }
                 } else {
                     return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[0]}` });
@@ -75,19 +77,19 @@ export class BrokerController {
             try {
                 const brokerService: BrokerService =
                     new BrokerService();
-    
+
                 if (Number(request.params.id)) {
                     const result: boolean =
                         await brokerService.alreadyRegisterById(Number(request.params.id));
-    
+
                     if (result) {
                         const result: boolean =
                             brokerService.validateData(request.body);
-    
+
                         if (result) {
                             const brokerEntity: BrokerEntity =
                                 await brokerService.update(Number(request.params.id), request.body, transaction);
-    
+
                             return response.status(200).json(brokerEntity);
                         } else {
                             return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[0]}` });
@@ -109,12 +111,35 @@ export class BrokerController {
             try {
                 const brokerService: BrokerService =
                     new BrokerService();
-    
+
                 if (Number(request.params.id)) {
-                    const deleteResult: DeleteResult =
-                        await brokerService.delete(Number(request.params.id), transaction);
-    
-                    return response.status(200).json({ manager: deleteResult.affected });
+                    const brokerEntity: BrokerEntity | undefined =
+                        await brokerService.read(Number(request.params.id));
+
+                    if (brokerEntity) {
+                        const userService: UserService =
+                            new UserService();
+
+                        const userEntity: UserEntity | undefined =
+                            await userService.findByBroker(Number(request.params.id));
+
+                        if (userEntity) {
+                            const userDeleteResult: DeleteResult =
+                                await userService.delete(userEntity.id, transaction);
+
+                            const brokerDeleteResult: DeleteResult =
+                                await brokerService.delete(Number(request.params.id), transaction);
+
+                            return response.status(200).json({
+                                user: userDeleteResult.affected,
+                                broker: brokerDeleteResult.affected
+                            });
+                        } else {
+                            return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[1]}` });
+                        }
+                    } else {
+                        return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[1]}` });
+                    }
                 } else {
                     return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[2]}` });
                 }
