+++ imob-web/back-end/src/api/controllers/manager.controller.ts
@@ -1,7 +1,9 @@
 import { DeleteResult, EntityManager, getManager } from 'typeorm';
 import { Request, Response } from 'express';
 import { ManagerEntity } from '../entities/manager.entity';
+import { UserEntity } from '../entities/user.entity';
 import { ManagerService } from '../services/manager.service';
+import { UserService } from '../services/user.service';
 import { returnMessages, statusMessages } from '../../../utils/utils';
 
 export class ManagerController {
@@ -27,21 +29,21 @@ export class ManagerController {
             try {
                 const managerService: ManagerService =
                     new ManagerService();
-    
+
                 const result: boolean =
                     managerService.validateData(request.body);
-    
+
                 if (result) {
                     const result: boolean =
                         await managerService.alreadyRegisterByCPF(request.body.cpf);
-    
+
                     if (!result) {
                         const managerEntity: ManagerEntity =
                             await managerService.create(request.body, transaction);
-    
+
                         return response.status(201).json(managerEntity);
                     } else {
-                        return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[3]}` });
+                        return response.status(409).json({ message: `${statusMessages[409]} ${returnMessages[3]}` });
                     }
                 } else {
                     return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[0]}` });
@@ -75,19 +77,19 @@ export class ManagerController {
             try {
                 const managerService: ManagerService =
                     new ManagerService();
-    
+
                 if (Number(request.params.id)) {
                     const result: boolean =
                         await managerService.alreadyRegisterById(Number(request.params.id));
-    
+
                     if (result) {
                         const result: boolean =
                             managerService.validateData(request.body);
-    
+
                         if (result) {
                             const managerEntity: ManagerEntity =
                                 await managerService.update(Number(request.params.id), request.body, transaction);
-    
+
                             return response.status(200).json(managerEntity);
                         } else {
                             return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[0]}` });
@@ -109,12 +111,35 @@ export class ManagerController {
             try {
                 const managerService: ManagerService =
                     new ManagerService();
-    
+
                 if (Number(request.params.id)) {
-                    const deleteResult: DeleteResult =
-                        await managerService.delete(Number(request.params.id), transaction);
-    
-                    return response.status(200).json({ manager: deleteResult.affected });
+                    const managerEntity: ManagerEntity | undefined =
+                        await managerService.read(Number(request.params.id));
+
+                    if (managerEntity) {
+                        const userService: UserService =
+                            new UserService();
+
+                        const userEntity: UserEntity | undefined =
+                            await userService.findByManager(Number(request.params.id));
+
+                        if (userEntity) {
+                            const userDeleteResult: DeleteResult =
+                                await userService.delete(userEntity.id, transaction);
+
+                            const managerDeleteResult: DeleteResult =
+                                await managerService.delete(Number(request.params.id), transaction);
+
+                            return response.status(200).json({
+                                user: userDeleteResult.affected,
+                                manager: managerDeleteResult.affected
+                            });
+                        } else {
+                            return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[1]}` });
+                        }
+                    } else {
+                        return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[1]}` });
+                    }
                 } else {
                     return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[2]}` });
                 }
