+++ imob-web/back-end/src/api/controllers/secretary.controller.ts
@@ -1,7 +1,9 @@
 import { DeleteResult, EntityManager, getManager } from 'typeorm';
 import { Request, Response } from 'express';
 import { SecretaryEntity } from '../entities/secretary.entity';
+import { UserEntity } from '../entities/user.entity';
 import { SecretaryService } from '../services/secretary.service';
+import { UserService } from '../services/user.service';
 import { returnMessages, statusMessages } from '../../../utils/utils';
 
 export class SecretaryController {
@@ -27,21 +29,21 @@ export class SecretaryController {
             try {
                 const secretaryService: SecretaryService =
                     new SecretaryService();
-    
+
                 const result: boolean =
                     secretaryService.validateData(request.body);
-    
+
                 if (result) {
                     const result: boolean =
                         await secretaryService.alreadyRegisterByCPF(request.body.cpf);
-    
+
                     if (!result) {
                         const secretaryEntity: SecretaryEntity =
                             await secretaryService.create(request.body, transaction);
-    
+
                         return response.status(201).json(secretaryEntity);
                     } else {
-                        return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[3]}` });
+                        return response.status(409).json({ message: `${statusMessages[409]} ${returnMessages[3]}` });
                     }
                 } else {
                     return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[0]}` });
@@ -75,19 +77,19 @@ export class SecretaryController {
             try {
                 const secretaryService: SecretaryService =
                     new SecretaryService();
-    
+
                 if (Number(request.params.id)) {
                     const result: boolean =
                         await secretaryService.alreadyRegisterById(Number(request.params.id));
-    
+
                     if (result) {
                         const result: boolean =
                             secretaryService.validateData(request.body);
-    
+
                         if (result) {
                             const secretaryEntity: SecretaryEntity =
                                 await secretaryService.update(Number(request.params.id), request.body, transaction);
-    
+
                             return response.status(200).json(secretaryEntity);
                         } else {
                             return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[0]}` });
@@ -109,12 +111,35 @@ export class SecretaryController {
             try {
                 const secretaryService: SecretaryService =
                     new SecretaryService();
-    
+
                 if (Number(request.params.id)) {
-                    const deleteResult: DeleteResult =
-                        await secretaryService.delete(Number(request.params.id), transaction);
-    
-                    return response.status(200).json({ manager: deleteResult.affected });
+                    const secretaryEntity: SecretaryEntity | undefined =
+                        await secretaryService.read(Number(request.params.id));
+
+                    if (secretaryEntity) {
+                        const userService: UserService =
+                            new UserService();
+
+                        const userEntity: UserEntity | undefined =
+                            await userService.findBySecretary(Number(request.params.id));
+
+                        if (userEntity) {
+                            const userDeleteResult: DeleteResult =
+                                await userService.delete(userEntity.id, transaction);
+
+                            const secretaryDeleteResult: DeleteResult =
+                                await secretaryService.delete(Number(request.params.id), transaction);
+
+                            return response.status(200).json({
+                                user: userDeleteResult.affected,
+                                secretary: secretaryDeleteResult.affected
+                            });
+                        } else {
+                            return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[1]}` });
+                        }
+                    } else {
+                        return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[1]}` });
+                    }
                 } else {
                     return response.status(400).json({ message: `${statusMessages[400]} ${returnMessages[2]}` });
                 }
